added risk notification component
