+++ App/screens/Home.tsx
@@ -13,8 +13,6 @@ import {RootStackParamList} from 'App/App';
 import Header from '../components/Header';
 
 import {global} from '../styles';
-import BasicButton from '../components/BasicButton';
-import {BlurBackground} from '../components/BackgroundBlur';
 import {useTranslation} from 'react-i18next';
 import {BottomMenu} from '../components/BottomMenu';
 import {ButtonPopup} from '../components/ButtonPopup';
@@ -137,8 +135,6 @@ export const Home: React.FC<{
 }> = ({navigation}) => {
   const {t} = useTranslation();
   const [distances, setDistances] = useState<never[]>([]);
-  const [showIDMatch, setIDMatchShow] = useState<boolean>(false);
-  const [hasSeenIDMatch, setIDMatchSeen] = useState<boolean>(false);
   const emitter = useRef<NativeEventEmitter | null>(null);
   const latestFetchTime = NativeModules.ItoBluetooth.getLatestFetchTime();
   console.log('Latest fetch time:', latestFetchTime);
@@ -163,21 +159,6 @@ export const Home: React.FC<{
       }
     };
   }, [distances.length]);
-
-  useEffect(() => {
-    function refresh(): void {
-      if (NativeModules.ItoBluetooth.isPossiblyInfected() && !hasSeenIDMatch) {
-        setIDMatchShow(true);
-      }
-    }
-    const interval = setInterval(refresh, 2500);
-    return (): void => clearInterval(interval);
-  }, [navigation, hasSeenIDMatch]);
-
-  const closeIDMatch = (): void => {
-    setIDMatchShow(false);
-    setIDMatchSeen(true);
-  };
 
   const closeDistances = distances.filter((d) => d <= 1.5);
   const furtherDistances = distances.filter((d) => d > 1.5 && d <= 5);
@@ -245,16 +226,6 @@ export const Home: React.FC<{
   return (
     <TouchableWithoutFeedback>
       <View style={[global.container, {paddingBottom: 100}]}>
-        {showIDMatch && (
-          <BlurBackground>
-            <View style={styles.IDMatchPopup}>
-              <Text style={styles.IDMatchText}>
-                {t('home.alertContactDiscovered')}
-              </Text>
-              <BasicButton title={t('home.whatNext')} onPress={closeIDMatch} />
-            </View>
-          </BlurBackground>
-        )}
         <Header />
         <View style={styles.lastFetchRow}>
           <Text style={styles.lastFetch}>
